moved zipFile to utils
